Added Text Labels for Each Data Line
This addition was more involved.  The text labels on each line of the
chart need to move with the data line as filters are applied onto the
data. To do this:

> I saved a copy of the d3 data globally so it can be accessed anywhere,
even after data has finished loading.
> I had to create a custom interpolation function (named interp1) to
determine the y-coordinate of the text label so it would always follow
the trend line.  It's easy to determine the x- value where the text
label is located, it is just at the last point in the x domain
(x.domain()). However, the tricky part is looking up the corresponding y
value for the x value. Interpolation was used, because sometimes the
chart line is cut off between data points, and we want the text label to
always be right next to the end of the line, not the nearest data point.
There may need to be future modifications to the interp1 function to
improve it's performance for large data sets.  A possible performance
increase could be added if interp1 remembers the two datapoints in the
dataset where it performed it's last interpolation, and started
searching for the new data points based on that instead of the beginning
of the data array.
> Changed margin formatting to make space for data labels.