Under the hood improvements - d3 simplification for creating multiple lines
Used the d3 tools to work on multiple nodes to render all of the lines
in the chart in one line of code. Each line function still needs to be
created, but then these are all added into an array that can be bound as
a data set to the path elements.  This way, when a new line is added to
the chart, only new liens need to be added into the lines array and
additional css added to format the color of each line.  There should no
longer be a need to add an extra line of code to render each line in the
d3.csv asynchronous call, and there should not be a need to add an extra
line of code to update the line in the brushed function as well.

> There may be some additional d3 simplifications that can be added as
well. Will investigate more tomorrow.