Got a line to render based on crossfilter
I'm saving the work so I can go back to the non-crossfitler graph and
compare to make sure the line is rendering the same as it was before.

Next steps will be to get the next line rendering, and the brush working
correctly again. There is also a problem with lining up the text label
for each line.  This is expected, as adding crossfilters is a major
restructuring that will require an architecture change of all these
components.