Got Text labels working - completed integration of crossfilter for data consolidation
Finished integrating crossfilters to consolidate the dataset.
> Text labels now work
> This chart now has the same functionality as the non-crossfilter
chart.
> Next steps:
1) Review code to implement crossfilter, and look for simplifications.
General clean up.  I just kind of hacked all of this together, and we
want to make sure we don't get too many hacks piled on top of each other
or else this code will become too convoluted and impossible to work
with.
2) Modify the time chart filter so it works through crossfilter, not
just rescaling the graph and chopping everything off with a clip path.