Reconstructed Line Chart
I only graphed number of stations per year, as this was an easier piece
of data to plot.

> To add multiple lines to the chart, you need to use a compositeChart
that allows multiple charts to be stacked.  Then, each line is a
seperate chart.
> crossfilter does not have a reduceAverage function, so this is one
reason I did not set it up to average data points. However, I was
reading that it's possible to construct a custom reduce average function
using crossfilters generic reduce() function.  There was an example on
the internet of doing this, and then using the reduction with dc.  It
was a stock market example, but it was something like this:
http://stackoverflow.com/questions/10509478/using-crossfilter-is-it-possible-to-track-max-min-when-grouping

Next step is to add a bar chart that can filter on capacity.