Bug Fixes and Code Cleanup
> Got bar bin sizes working.
> Reorganized code, and removed additional unnecessary code
> Other minor fixes
> The new excel file labeled w Pivot (maybe also in previous change),
has excel graphs that match the graphs in this webpage.  This was just a
tool to check that the results being displayed were accurate.

With more work, I think it would be possible to graph two lines of
average $/kw.  There is a compoundChart type in dc that allows you to
overlay charts, which is how you can get multiple lines into a chart.
Also, to do a reduceAverage in crossfilter, you need to build a custom
reduce funciton using reduce().  I read some stuff and saw an example on
the internet of how to do a reduce Average, might take a day to figure
out.

>