Attempt to add reduce by average graph
The no incentive line appears to be rendering properly, as it appears to
have the same shape as the old versions. However, the y-axis labels are
not rendering properly, so this needs to be fixed.

Testing is needed to weed out any bugs in the crossfilter render
function.  For example, currently there will be a divide by 0 error if
the count reaches 0.